requirements.txt
asn1crypto==0.24.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.17.1
idna==2.8
mysql-connector==2.2.9
numpy==1.17.2
opencv-python==4.1.1.26
pycparser==2.19
python-dotenv==0.10.3
python-telegram-bot==12.1.0
requests==2.22.0
six==1.12.0
tornado==6.0.3
urllib3==1.25.3
